refactor: enhance useNextToActInfo hook with improved type safety and remove unnecessary memoization
