+++ ui/src/hooks/useNextToActInfo.ts
@@ -1,23 +1,14 @@
 import { useGameState } from "./useGameState";
-import { useCallback, useMemo } from "react";
-import { PlayerDTO } from "@bitcoinbrisbane/block52";
-import useSWR from "swr";
-
-// Define the nextToActInfo type
-export interface NextToActInfo {
-    seat: number;
-    player: PlayerDTO;
-    isCurrentUserTurn: boolean;
-    availableActions: any[];
-    timeRemaining: number;
-}
+import { useMemo } from "react";
+import { PlayerDTO, LegalActionDTO, TexasHoldemStateDTO } from "@bitcoinbrisbane/block52";
+import { NextToActInfo, NextToActInfoReturn, GameStateReturn } from "../types/index";
 
 /**
  * Determines who is next to act at the table
  * @param gameData Current game state data
  * @returns Object containing information about who is next to act
  */
-export const whoIsNextToAct = (gameData: any): NextToActInfo | null => {
+export const whoIsNextToAct = (gameData: TexasHoldemStateDTO): NextToActInfo | null => {
     if (!gameData || !gameData.players) return null;
 
     const nextToActSeat = gameData.nextToAct;
@@ -31,8 +22,8 @@ export const whoIsNextToAct = (gameData: any): NextToActInfo | null => {
     const userAddress = localStorage.getItem("user_eth_public_key")?.toLowerCase();
     const isCurrentUserTurn = player.address?.toLowerCase() === userAddress;
 
-    // Get available actions
-    const availableActions = player.legalActions || [];
+    // Get available actions (properly typed as LegalActionDTO[])
+    const availableActions: LegalActionDTO[] = player.legalActions || [];
 
     // Calculate time remaining (if needed)
     const timeRemaining = player.timeout || 30; // Default to 30 seconds
@@ -51,20 +42,9 @@ export const whoIsNextToAct = (gameData: any): NextToActInfo | null => {
  * @param tableId The ID of the table to fetch state for
  * @returns Object containing next-to-act information
  */
-export const useNextToActInfo = (tableId?: string) => {
-    // Memoize the lastRefresh reference to avoid unnecessary rerenders
-    const lastRefreshRef = useCallback(() => {
-        let lastTime = 0;
-        return {
-            get: () => lastTime,
-            set: (time: number) => {
-                lastTime = time;
-            }
-        };
-    }, [])();
-
+export const useNextToActInfo = (tableId?: string): NextToActInfoReturn => {
     // Get game state from centralized hook
-    const { gameState, isLoading, error, refresh } = useGameState(tableId);
+    const { gameState, isLoading, error, refresh }: GameStateReturn = useGameState(tableId);
 
     // Memoize the nextToActInfo calculation to avoid recomputing on every render
     const nextToActInfo = useMemo(() => {
@@ -85,40 +65,11 @@ export const useNextToActInfo = (tableId?: string) => {
         }
     }, [gameState, isLoading, error]);
 
-    // Memoize the fetcher function to avoid recreation on every render
-    const fetcher = useCallback(async () => {
-        const now = Date.now();
-        // Refresh if more than 3 seconds have elapsed
-        if (now - lastRefreshRef.get() >= 3000) {
-            await refresh();
-            lastRefreshRef.set(now);
-        }
-        return null;
-    }, [refresh, lastRefreshRef]);
-
-    // Custom more frequent refresh using SWR
-    useSWR(tableId ? `next-to-act-${tableId}` : null, fetcher, {
-        refreshInterval: 3000,
-        revalidateOnFocus: true,
-        // Reduce unnecessary revalidations
-        dedupingInterval: 1000,
-        // Prevent additional renders from focus events when tab switching
-        focusThrottleInterval: 5000
-    });
-
-    // Memoize the manual refresh function to maintain reference equality
-    const manualRefresh = useCallback(() => {
-        return refresh();
-    }, [refresh]);
-
-    // Memoize the result object to prevent unnecessary re-renders in consuming components
-    return useMemo(
-        () => ({
-            nextToActInfo,
-            isLoading,
-            error,
-            refresh: manualRefresh
-        }),
-        [nextToActInfo, isLoading, error, manualRefresh]
-    );
+    // Return the result object
+    return {
+        nextToActInfo,
+        isLoading,
+        error,
+        refresh
+    };
 };
